requirements.txt
paho-mqtt
Flask
Flask-Cors
mysql-connector-python
streamlit
pandas
python-dotenv
